requirements.txt
feedparser==6.0.11
requests==2.32.3
pydantic==2.8.2
python-dateutil==2.9.0.post0
pyyaml>=6.0

